+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditor.java
@@ -43,7 +43,6 @@ import javax.swing.*;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import javax.swing.text.JTextComponent;
-
 import net.sf.jabref.*;
 import net.sf.jabref.gui.actions.Actions;
 import net.sf.jabref.gui.fieldeditors.*;
@@ -51,6 +50,7 @@ import net.sf.jabref.gui.keyboard.KeyBinds;
 import net.sf.jabref.bibtex.BibtexEntryWriter;
 import net.sf.jabref.gui.menus.ChangeEntryTypeMenu;
 import net.sf.jabref.logic.autocompleter.AutoCompleter;
+import net.sf.jabref.logic.journals.Abbreviations;
 import net.sf.jabref.exporter.LatexFieldFormatter;
 import net.sf.jabref.external.ExternalFilePanel;
 import net.sf.jabref.external.WriteXMPEntryEditorAction;
@@ -59,13 +59,11 @@ import net.sf.jabref.gui.date.DatePickerButton;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.importer.fileformat.BibtexParser;
 import net.sf.jabref.importer.ParserResult;
-import net.sf.jabref.gui.journals.JournalAbbreviationsUtil;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.labelPattern.LabelPatternUtil;
 import net.sf.jabref.logic.util.date.EasyDateFormat;
 import net.sf.jabref.model.database.BibtexDatabase;
 import net.sf.jabref.model.entry.*;
-import net.sf.jabref.bibtex.EntryTypes;
 import net.sf.jabref.specialfields.SpecialFieldUpdateListener;
 import net.sf.jabref.gui.undo.NamedCompound;
 import net.sf.jabref.gui.undo.UndoableChangeType;
@@ -170,6 +168,14 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
 
     private final TabListener tabListener = new TabListener();
 
+    // @formatter:off
+    private final String ABBREVIATION_TOOLTIP_TEXT = "<HTML>"
+            + Localization.lang("Switches between full and abbreviated journal name if the journal name is known.")
+            + "<BR>" + Localization.lang("To set up, go to") + " <B>"
+            + Localization.lang("Options") + " -> "
+            + Localization.lang("Manage journal abbreviations") + "</B></HTML>";
+    // @formatter:on
+
 
     public EntryEditor(JabRefFrame frame, BasePanel panel, BibtexEntry entry) {
 
@@ -205,7 +211,7 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     private void setupFieldPanels() {
         tabbed.removeAll();
         tabs.clear();
-        List<String> fieldList = entry.getRequiredFields();
+        List<String> fieldList = entry.getRequiredFieldsFlat();
 
         EntryEditorTab reqPan = new EntryEditorTab(frame, panel, fieldList, this, true, false, Localization.lang("Required fields"));
         if (reqPan.fileListEditor != null) {
@@ -423,8 +429,8 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
         String fieldExtras = BibtexFields.getFieldExtras(string);
 
         // timestamp or a other field with datepicker command
-        if (fieldName.equals(Globals.prefs.get(JabRefPreferences.TIME_STAMP_FIELD))
-                || ((fieldExtras != null) && fieldExtras.equals("datepicker"))) {
+        if (Globals.prefs.get(JabRefPreferences.TIME_STAMP_FIELD).equals(fieldName)
+                || BibtexFields.EXTRA_DATEPICKER.equals(fieldExtras)) {
             // double click AND datefield => insert the current date (today)
             ((JTextArea) editor).addMouseListener(new MouseAdapter() {
 
@@ -439,19 +445,19 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
             });
 
             // insert a datepicker, if the extras field contains this command
-            if ((fieldExtras != null) && fieldExtras.equals("datepicker")) {
+            if (BibtexFields.EXTRA_DATEPICKER.equals(fieldExtras)) {
                 DatePickerButton datePicker = new DatePickerButton(editor);
                 return datePicker.getDatePicker();
             }
         }
 
-        if ((fieldExtras != null) && fieldExtras.equals("external")) {
+        if (BibtexFields.EXTRA_EXTERNAL.equals(fieldExtras)) {
 
             // Add external viewer listener for "pdf" and "url" fields.
             ((JComponent) editor).addMouseListener(new ExternalViewerListener());
 
             return null;
-        } else if ((fieldExtras != null) && fieldExtras.equals("journalNames")) {
+        } else if (BibtexFields.EXTRA_JOURNAL_NAMES.equals(fieldExtras)) {
             // Add controls for switching between abbreviated and full journal
             // names.
             // If this field also has a FieldContentSelector, we need to combine
@@ -464,8 +470,29 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
                 contentSelectors.add(ws);
                 controls.add(ws, BorderLayout.NORTH);
             }
-            controls.add(JournalAbbreviationsUtil.getNameSwitcher(this, editor, panel.undoManager),
-                    BorderLayout.SOUTH);
+
+            // Button to toggle abbreviated/full journal names
+            JButton button = new JButton(Localization.lang("Toggle abbreviation"));
+            button.setToolTipText(ABBREVIATION_TOOLTIP_TEXT);
+            button.addActionListener(new ActionListener() {
+
+                @Override
+                public void actionPerformed(ActionEvent actionEvent) {
+                    String text = editor.getText();
+                    if (Abbreviations.journalAbbrev.isKnownName(text)) {
+                        String s = Abbreviations.toggleAbbreviation(text);
+
+                        if (s != null) {
+                            editor.setText(s);
+                            storeFieldAction.actionPerformed(new ActionEvent(editor, 0, ""));
+                            panel.undoManager
+                                    .addEdit(new UndoableFieldChange(getEntry(), editor.getFieldName(), text, s));
+                        }
+                    }
+                }
+            });
+
+            controls.add(button, BorderLayout.SOUTH);
             return controls;
         } else {
             if (panel.metaData.getData(Globals.SELECTOR_META_PREFIX + editor.getFieldName()) != null) {
@@ -476,7 +503,7 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
 
                 return ws;
             } else {
-                if ((fieldExtras != null) && fieldExtras.equals("browse")) {
+                if (BibtexFields.EXTRA_BROWSE.equals(fieldExtras)) {
                     JButton but = new JButton(Localization.lang("Browse"));
                     ((JComponent) editor).addMouseListener(new ExternalViewerListener());
 
@@ -504,11 +531,12 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
 
                     return but;
 
-                } else if ((fieldExtras != null) && (fieldExtras.equals("browseDoc") || fieldExtras.equals("browseDocZip"))) {
+                } else if (BibtexFields.EXTRA_BROWSE_DOC.equals(fieldExtras)
+                        || BibtexFields.EXTRA_BROWSE_DOC_ZIP.equals(fieldExtras)) {
 
                     final String ext = '.' + fieldName.toLowerCase();
                     final OpenFileFilter off;
-                    if (fieldExtras.equals("browseDocZip")) {
+                    if (BibtexFields.EXTRA_BROWSE_DOC_ZIP.equals(fieldExtras)) {
                         off = new OpenFileFilter(new String[]{ext, ext + ".gz", ext + ".bz2"});
                     } else {
                         off = new OpenFileFilter(new String[]{ext});
@@ -516,12 +544,12 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
 
                     return new ExternalFilePanel(frame, panel.metaData(), this, fieldName,
                             off, editor);
-                } else if ((fieldExtras != null) && fieldExtras.equals("url")) {
+                } else if (BibtexFields.EXTRA_URL.equals(fieldExtras)) {
                     ((JComponent) editor).setDropTarget(new DropTarget((Component) editor,
                             DnDConstants.ACTION_NONE, new SimpleUrlDragDrop(editor, storeFieldAction)));
 
                     return null;
-                } else if ((fieldExtras != null) && fieldExtras.equals("setOwner")) {
+                } else if (BibtexFields.EXTRA_SET_OWNER.equals(fieldExtras)) {
                     JButton button = new JButton(Localization.lang("Auto"));
                     button.addActionListener(new ActionListener() {
 
@@ -532,6 +560,45 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
                         }
                     });
                     return button;
+                } else if (BibtexFields.EXTRA_YES_NO.equals(fieldExtras)) {
+                    final String[] options = {"", "Yes", "No"};
+                    JComboBox<String> yesno = new JComboBox<>(options);
+                    yesno.addActionListener(new ActionListener() {
+
+                        @Override
+                        public void actionPerformed(ActionEvent actionEvent) {
+
+                            editor.setText(((String) yesno.getSelectedItem()).toLowerCase());
+                            updateField(editor);
+                        }
+                    });
+                    return yesno;
+                } else if (BibtexFields.EXTRA_MONTH.equals(fieldExtras)) {
+                    final String[] options = new String[13];
+                    options[0] = Localization.lang("Select");
+                    for (int i = 1; i <= 12; i++) {
+                        options[i] = MonthUtil.getMonthByNumber(i).fullName;
+                    }
+                    JComboBox<String> month = new JComboBox<>(options);
+                    month.addActionListener(new ActionListener() {
+
+                        @Override
+                        public void actionPerformed(ActionEvent actionEvent) {
+                            int monthnumber = month.getSelectedIndex();
+                            if (monthnumber >= 1) {
+                                if (prefs.getBoolean(JabRefPreferences.BIBLATEX_MODE)) {
+                                    editor.setText(String.valueOf(monthnumber));
+                                } else {
+                                    editor.setText((MonthUtil.getMonthByNumber(monthnumber).bibtexFormat));
+                                }
+                            } else {
+                                editor.setText("");
+                            }
+                            updateField(editor);
+                            month.setSelectedIndex(0);
+                        }
+                    });
+                    return month;
                 } else {
                     return null;
                 }
@@ -1030,9 +1097,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     private class TypeLabel extends JLabel {
-
-        private static final long serialVersionUID = 1L;
-
         public TypeLabel(String type) {
             super(type);
             setUI(new VerticalLabelUI(false));
@@ -1123,9 +1187,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class DeleteAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public DeleteAction() {
             super(Localization.lang("Delete"), IconTheme.JabRefIcon.DELETE.getIcon());
             putValue(Action.SHORT_DESCRIPTION, Localization.lang("Delete entry"));
@@ -1149,9 +1210,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class CloseAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public CloseAction() {
             super(Localization.lang("Close window"), IconTheme.JabRefIcon.CLOSE.getSmallIcon());
             putValue(Action.SHORT_DESCRIPTION, Localization.lang("Close window"));
@@ -1171,9 +1229,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     public class StoreFieldAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public StoreFieldAction() {
             super("Store field value");
             putValue(Action.SHORT_DESCRIPTION, "Store field value");
@@ -1336,9 +1391,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class SwitchLeftAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public SwitchLeftAction() {
             super("Switch to the panel to the left");
         }
@@ -1353,9 +1405,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class SwitchRightAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public SwitchRightAction() {
             super("Switch to the panel to the right");
         }
@@ -1370,9 +1419,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class NextEntryAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public NextEntryAction() {
             super(Localization.lang("Next entry"), IconTheme.JabRefIcon.DOWN.getIcon());
 
@@ -1401,9 +1447,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class PrevEntryAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public PrevEntryAction() {
             super(Localization.lang("Previous entry"), IconTheme.JabRefIcon.UP.getIcon());
 
@@ -1432,11 +1475,7 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class GenerateKeyAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         final JabRefFrame parent;
-
 
         public GenerateKeyAction(JabRefFrame parentFrame) {
             super(Localization.lang("Generate BibTeX key"), IconTheme.JabRefIcon.MAKE_KEY.getIcon());
@@ -1497,9 +1536,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class UndoAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public UndoAction() {
             super("Undo", IconTheme.JabRefIcon.UNDO.getIcon());
             putValue(Action.SHORT_DESCRIPTION, "Undo");
@@ -1512,9 +1548,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class RedoAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public RedoAction() {
             super("Redo", IconTheme.JabRefIcon.REDO.getIcon());
             putValue(Action.SHORT_DESCRIPTION, "Redo");
@@ -1527,9 +1560,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
     }
 
     class SaveDatabaseAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public SaveDatabaseAction() {
             super("Save database");
         }
@@ -1608,9 +1638,6 @@ public class EntryEditor extends JPanel implements VetoableChangeListener, Entry
 
 
     private class AutoLinkAction extends AbstractAction {
-
-        private static final long serialVersionUID = 1L;
-
         public AutoLinkAction() {
             putValue(Action.SMALL_ICON, IconTheme.JabRefIcon.AUTO_FILE_LINK.getIcon());
             putValue(Action.SHORT_DESCRIPTION, Localization.lang("Automatically set file links for this entry") + " (Alt-F)");