+++ src/main/java/net/sf/jabref/model/entry/EntryUtil.java
@@ -1,6 +1,7 @@
 package net.sf.jabref.model.entry;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public class EntryUtil {
@@ -16,11 +17,14 @@ public class EntryUtil {
         return one == null ? two == null : one.equals(two);
     }
 
+    /**
+     * Make first character of String uppercase, and the
+     * rest lowercase.
+     */
     public static String capitalizeFirst(String toCapitalize) {
-        // Make first character of String uppercase, and the
-        // rest lowercase.
         if (toCapitalize.length() > 1) {
-            return toCapitalize.substring(0, 1).toUpperCase() + toCapitalize.substring(1, toCapitalize.length()).toLowerCase();
+            return toCapitalize.substring(0, 1).toUpperCase()
+                    + toCapitalize.substring(1, toCapitalize.length()).toLowerCase();
         } else {
             return toCapitalize.toUpperCase();
         }
@@ -30,17 +34,19 @@ public class EntryUtil {
     /**
      * Build a String array containing all those elements of all that are not
      * in subset.
+     *
      * @param all The array of all values.
      * @param subset The subset of values.
-     * @return The remainder that is not part of the subset.
+     * @return The remainder that is not part of the subset. - The result MUST NOT be modified
      */
     public static List<String> getRemainder(List<String> all, List<String> subset) {
-    	if (subset.size() == 0) {
-    		return all;
-    	}
-    	if (all.equals(subset)) {
-    		return new ArrayList<>();
-    	}
+        if (subset.isEmpty()) {
+            // ensure that "all" does not get modified
+            return Collections.unmodifiableList(all);
+        }
+        if (all.equals(subset)) {
+            return Collections.EMPTY_LIST;
+        }
 
         ArrayList<String> al = new ArrayList<>();
         for (String anAll : all) {