+++ src/main/java/net/sf/jabref/model/entry/IEEETranEntryTypes.java
@@ -1,9 +1,6 @@
 package net.sf.jabref.model.entry;
 
-import net.sf.jabref.model.database.BibtexDatabase;
-
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 /**
@@ -13,44 +10,50 @@ import java.util.List;
  * Electronic, IEEETranBSTCTL, Periodical, Patent, Standard
  */
 public class IEEETranEntryTypes {
+
     /**
      * Electronic entry type for internet references
      *
      * Required fields:
      * Optional fields: author, month, year, title, language, howpublished, organization, address, note, url
      */
-    public static final EntryType ELECTRONIC =
-            new BibtexEntryType() {
+    public static final EntryType ELECTRONIC = new BibtexEntryType() {
 
-                {
-                    addAllOptional("author", "month", "year", "title", "language", "howpublished", "organization", "address", "note", "url");
+        {
+            addAllOptional("author", "month", "year", "title", "language", "howpublished", "organization", "address",
+                    "note", "url");
 
-                }
+        }
 
-                @Override
-                public String getName() {
-                    return "Electronic";
-                }
-            };
+        @Override
+        public String getName() {
+            return "Electronic";
+        }
+    };
 
     /**
      * Special entry type that can be used to externally control some aspects of the bibliography style.
      */
-    public static final EntryType IEEETRANBSTCTL =
-            new BibtexEntryType() {
+    public static final EntryType IEEETRANBSTCTL = new BibtexEntryType() {
 
-                {
-                    addAllOptional("ctluse_article_number", "ctluse_paper", "ctluse_forced_etal", "ctluse_url",
-                            "ctlmax_names_forced_etal", "ctlnames_show_etal", "ctluse_alt_spacing",
-                            "ctlalt_stretch_factor", "ctldash_repeated_names", "ctlname_format_string",
-                            "ctlname_latex_cmd", "ctlname_url_prefix");
-                }
+        {
+            addAllOptional("ctluse_article_number", "ctluse_paper", "ctluse_forced_etal", "ctluse_url",
+                    "ctlmax_names_forced_etal", "ctlnames_show_etal", "ctluse_alt_spacing", "ctlalt_stretch_factor",
+                    "ctldash_repeated_names", "ctlname_format_string", "ctlname_latex_cmd", "ctlname_url_prefix");
+        }
 
-                @Override
-                public String getName() {
-                    return "IEEEtranBSTCTL";
-                }
-            };
+        @Override
+        public String getName() {
+            return "IEEEtranBSTCTL";
+        }
+
+    };
+
+    public static final String[] IEEETRANBSTCTL_YES_NO_FIELDS = {"ctluse_article_number", "ctluse_paper", "ctluse_url",
+            "ctluseforced_etal", "ctluse_alt_spacing", "ctldash_repeated_names"};
+
+    public static final String[] IEEETRANBSTCTL_NUMERIC_FIELDS = {"ctlmax_names_forced_etal", "ctlnames_show_etal",
+            "ctlalt_stretch_factor"};
 
     /**
      * The periodical entry type is used for journals and magazines.
@@ -58,19 +61,18 @@ public class IEEETranEntryTypes {
      * Required fields: title, year
      * Optional fields: editor, language, series, volume, number, organization, month, note, url
      */
-    public static final EntryType PERIODICAL =
-            new BibtexEntryType() {
+    public static final EntryType PERIODICAL = new BibtexEntryType() {
 
-                {
-                    addAllOptional("editor", "language", "series", "volume", "number", "organization", "month", "note", "url");
-                    addAllRequired("title", "year");
-                }
+        {
+            addAllRequired("title", "year");
+            addAllOptional("editor", "language", "series", "volume", "number", "organization", "month", "note", "url");
+        }
 
-                @Override
-                public String getName() {
-                    return "Periodical";
-                }
-            };
+        @Override
+        public String getName() {
+            return "Periodical";
+        }
+    };
 
     /**
      * Entry type for patents.
@@ -78,19 +80,19 @@ public class IEEETranEntryTypes {
      * Required fields: nationality, number, year or yearfiled
      * Optional fields: author, title, language, assignee, address, type, number, day, dayfiled, month, monthfiled, note, url
      */
-    public static final EntryType PATENT =
-            new BibtexEntryType() {
+    public static final EntryType PATENT = new BibtexEntryType() {
 
-                {
-                    addAllOptional("author", "title", "language", "assignee", "address", "type", "number", "day", "dayfiled", "month", "monthfiled", "note", "url");
-                    addAllRequired("nationality", "number", "year/yearfiled");
-                }
+        {
+            addAllRequired("nationality", "number", "year/yearfiled");
+            addAllOptional("author", "title", "language", "assignee", "address", "type", "number", "day", "dayfiled",
+                    "month", "monthfiled", "note", "url");
+        }
 
-                @Override
-                public String getName() {
-                    return "Patent";
-                }
-            };
+        @Override
+        public String getName() {
+            return "Patent";
+        }
+    };
 
     /**
      * The standard entry type is used for proposed or formally published standards.
@@ -98,18 +100,19 @@ public class IEEETranEntryTypes {
      * Required fields: title, organization or institution
      * Optional fields: author, language, howpublished, type, number, revision, address, month, year, note, url
      */
-    public static final EntryType STANDARD =
-            new BibtexEntryType() {
-                {
-                    addAllOptional("author", "language", "howpublished", "type", "number", "revision", "address", "month", "year", "note", "url");
-                    addAllRequired("title", "organization/institution");
-                }
+    public static final EntryType STANDARD = new BibtexEntryType() {
 
-                @Override
-                public String getName() {
-                    return "Standard";
-                }
-            };
+        {
+            addAllRequired("title", "organization/institution");
+            addAllOptional("author", "language", "howpublished", "type", "number", "revision", "address", "month",
+                    "year", "note", "url");
+        }
+
+        @Override
+        public String getName() {
+            return "Standard";
+        }
+    };
 
     public static final List<EntryType> ALL = Arrays.asList(ELECTRONIC, IEEETRANBSTCTL, PERIODICAL, PATENT, STANDARD);
 }